fix(training): allow for string in AnemoiDatasetsMask.get_grid_size  (#604)
## Description
Some configurations pass a dataset path as a string to the
`AnemoiDatasetsMask.get_grid_size` through `nodes['_dataset']` instead
of a dictionary. Since I'm not sure how to change this behavior on the
configuration I have updated the code to accept both str and dict types
as it is still compatible with open_dataset.

Co-authored-by: Mario Santa Cruz <48736305+JPXKQX@users.noreply.github.com>