feat(models): point-mlp proccesor (#367)
Currently, the land component of the IFS (ecLand) is emulated using a
simple MLP coded outside of Anemoi-core.

To be able to utilise the full Anemoi workflow when emulating the land
surface, we are adding a feature that would simplify the training in the
anemoi-core, allowing for MLP-like training. This means effectively
by-passing the graph structure by letting points be their own
neighbours. This allows us to keep the single column structure of ecLand
(no latent information between grids cells) and have a resolution
agnostic model (i.e., if we train the model weights on an o96 dataset we
can directly apply this to n320, as long as trainable_parameters is set
to 0).

## Description
<!-- What issue or task does this change relate to? -->
This PR brings the following new functionality.
- A new processor mapper, ``PointWiseMLPProcessor``, to processes each
node independently using a MLP.
- Tests have been added for the new processor, block and chunk.
- Schemas added for processors.


<!-- readthedocs-preview anemoi-training start -->
----
📚 Documentation preview 📚:
https://anemoi-training--367.org.readthedocs.build/en/367/

<!-- readthedocs-preview anemoi-training end -->

<!-- readthedocs-preview anemoi-graphs start -->
----
📚 Documentation preview 📚:
https://anemoi-graphs--367.org.readthedocs.build/en/367/

<!-- readthedocs-preview anemoi-graphs end -->

<!-- readthedocs-preview anemoi-models start -->
----
📚 Documentation preview 📚:
https://anemoi-models--367.org.readthedocs.build/en/367/

<!-- readthedocs-preview anemoi-models end -->

---------

Co-authored-by: Mario Santa Cruz <mariosanta_cruz@hotmail.com>
Co-authored-by: Ewan Pinnington <ewan.pinnington@ecmwf.int>
Co-authored-by: anaprietonem <ana.prietonemesio@ecmwf.int>