fix(models): replace placeholder of migration script anemoi-models version in main (#600)
## Description
Before a new version is published, the anemoi-models version stored in
migration scripts is set as a placeholder
`%NEXT_ANEMOI_MODELS_VERSION%`. If the migrations are run on the main
branch with script still having the placeholder, the checkpoint will
incorrect information of the version of the scripts.

This replaces the placeholder with the current `__version__` of
anemoi-models (as the future version is impossible to predict).

This will allow a better feedback by `anemoi-models migration inspect`. 

## What problem does this change solve?
<!-- Describe if it's a bugfix, new feature, doc update, or breaking
change -->

## What issue or task does this change relate to?
<!-- link to Issue Number -->

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

Co-authored-by: Ana Prieto Nemesio <91897203+anaprietonem@users.noreply.github.com>