fix(graphs): skip expensive graph creation when target exists without overwrite (#607)
## Description
This PR refactors the `anemoi-graphs create` workflow to avoid redundant
graph creation when a target graph file already exists and the
--overwrite flag is not set.

The previous check in the `save` function of GraphCreator has been
preserved for concurrency protection, but has had its error message
refined.

## What problem does this change solve?
Currently, this check is only evaluated at save time, so the whole graph
creation process is still performed even though the graph won't be
saved.

Additionally, the previous error message was hidden in the middle of the
output log, whereas now it is prominent at the top.


## What issue or task does this change relate to?
no existing issue
##  Additional notes ##
- Some small additional tweaks due to `black` formatting.

- Maybe it's not a good idea to print the description of the existing
graph in the case of a failed overwrite? \
 (ie drop the `GraphDescriptor(args.save_path).describe()`)

It's slightly confusing for a process that is supposed to create
something new, but fails instead—it can make users fail to notice the
failure (as happened to me).
   
   Currently, it's been preserved for compatibility reasons.

- Should this also be an `ERROR`, not just a `WARNING`? 

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Ana Prieto Nemesio <91897203+anaprietonem@users.noreply.github.com>