fix(graphs): unit-range normalisation (#665)
## Description
<!-- What issue or task does this change relate to? -->
This PR fixes a bug in `anemoi-graphs` when using the following
normalisation method:
```
norm: unit-range
```
This is the only method that returns 2 statistics, and it should use the
second argument (-1, the range) to scale the attributes.


This PR only change the behaviour when `norm: unit-range`, as in the
other methods, `statistics[0] = statistics[-1]`.


***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

Co-authored-by: Ana Prieto Nemesio <91897203+anaprietonem@users.noreply.github.com>