fix(training): provide more informative error when user specifies inexistent node attribute  (#663)
Small fix to make the error message more informative. Instead of just
telling a user that the specified attribute does not exist in the graph
data, it also provides a list of the ones that are available. So for
instance, if one mistakenly specified "cutout_mask" instead of "cutout",
they will know right away what the problem is.