feat(models): multibackend all_to_all wrapper (#95)
Small addition to fallback to support alltoall when using the Gloo backend to
torch.distributed. This PR is needed to be able to run the transformer
model on CPU. for 99.9% of users running on GPUs with the NCCL
background, this change should not effect them

Gloo does not offer an alltoall primitive, as shown
[here](https://pytorch.org/docs/stable/distributed.html#backends)

This commit implements am all_to_all fallback for Gloo, using the 'Linear
Shift' algorithm from [Hoffman and Rünger,
2013](https://www.tu-chemnitz.de/informatik/PI/forschung/publikationen/download/HR_eurompi13.pdf).
Because of syntax for `torch.dist` changing in torch 2.6, older versions
of torch are not supported.

---------

Co-authored-by: Harrison Cook <Harrison.cook@ecmwf.int>